+++ tests/adapters_unit_tests/controllers/book_controller_tests.cpp
@@ -58,7 +58,7 @@ public:
     MOCK_METHOD(void, setupUserData, (const QString&, const QString&),
                 (override));
     MOCK_METHOD(void, clearUserData, (), (override));
-    MOCK_METHOD(void, updateUsedBookStorage, (double), (override));
+    MOCK_METHOD(void, updateUsedBookStorage, (long), (override));
 };
 
 struct ABookController : public ::testing::Test
